Refactor file structure to allow for separate style.js file
