+++ client/src/Components/index.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App.jsx';
+import App from './App';
 
 ReactDOM.render(<App />, document.getElementById('app'));
