+++ webpack.config.js
@@ -1,6 +1,6 @@
 const path = require('path');
 
-const SRC_DIR = path.join(__dirname, '/client/src');
+const SRC_DIR = path.join(__dirname, '/client/src/Components');
 const DIST_DIR = path.join(__dirname, '/public');
 
 module.exports = {
